// types/schema.d.ts
export interface components {
    schemas: {
// ...
        UserResponse: {
            /** Member Id */
            member_id: string;
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Full Name */
            full_name: string;
            /** Phone */
            phone?: string | null;
            /** @default member */
            role: components["schemas"]["UserRole"];
            /** Id */
            id: number;
            status: components["schemas"]["UserStatus"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}